Documentation fix.
To be more precise: it should be "the last event whose every property
except for its value matchod that same KEY", but that wording is
difficult to understand. Type, code and domain should be the only
properties besides value that can currently be specified for hooks,
so the new wording should be accurate.