Behaviour change: patched --toggle mode=consistent
Consistent-mode toggles will now lock event targets based on
(type, code, domain) tuples, rather than (type, code) pairs.

This is strictly speaking a breaking change. The documention
mentions that it will remember "for every key" without defining
what key means. Formerly that was a (type, code) pair with type=EV_KEY,
now it has been reinterpreted to mean a (type, code, domain) tuple.

In the very few cases where the behaviour matters, it is however hard
to see the consequences of the old behaviour as anything but a bug.

There are very few scripts will be affected by this. After weighing
the derire of having "stable behaviour" against the desire to have
"bug-free behaviour", I decided that bug-free behaviour wins in
this case.