Block (almost) all signals except during Epoll::poll_raw().
Signals can cause some syscalls to fail for virtually no reason other
than "Interrupt received during this call", which can confuse evsieve
and make it think that something more fundamental is wrong such as
having a broken file descriptor. To reduce the risk of (part of) evsieve
crashing due to receiving a benign signal that breaks a syscall, all
signals except SIGSEGV will now be delayed until handled by
Epoll::poll_raw(), which is designed to be able to deal with benign
interruptions.