Started migrating away from State-based mutability.
Although the State class is a useful way to satisfy the Rust borrow
checker without having to resort to internal mutability, it does cause
quite a bit of cognitive overhead. In the end, I think that using the
State class was a bad idea that needlessly increases code complexity for
little reward. I am now stating to refactor the code base to cease usage
of the State altogether and instead use internal mutability like Cell
and Rc<RefCell> instead.