Performance optimisation.
Somehow the mere possibility of printing a warning message for a
violation of an internal variant during a tight inner loop will make
evsieve's performance 33% slower on some specific benchmarks. This
is probably because it affects the compiler/CPU's ability to reorder
certain operations.

Having to choose between "33% worse performance of the input system"
and "a warning might not be generated in a situation that should
never happen anyway", the latter seems to be the lesser evil.