Revert the last three commits.
It turns out that the poll() interface is not actually more convenient
than the Epoll interface at some parts of the program, and I would
rather not have two different interfaces used at the same time, so I am
reverting the shift to poll(). I'll keep those commits in the git
history just in case I change my mind in the future though.