Rename relative keys like "rel:x:2delta" to "rel:x:2d"
My inner mathematician says that there must either be a multiplication
symbol like "2*delta", or the variables must contain only a single
letter like "2d". Because * is a character that is interpreted by the
shell, using that as part of an argument is a landmine, therefore the
latter approach has been chosen.