Revert dynamically loading libsystemd at runtime.
Dynamically loading systemd does not solve the issue of needing the
systemd development files installed. I could try to do a hybrid solution
"try to link against systemd at compile time, if not link against it at
runtime, if that fails fall back on systemd-notify." but that creates a
complex flowchart on when sd_notify will or will not fail, which creates
a maintainance burden. Maybe in the future I'll re-add dynamically
linking, but for now I'd like to avoid creating that maintainance burden
with the 1.3 release, so I will use "rely on systemd-notify" exclusively
and add its known caveats to the README.