Updated benchmark results.
The previous benchmark was measured by a Python script that measured
the time between writing events to a virtual input device and reading
them back from from the virtual output device. This benchmark was flawed
because the time it took the Python script to write and read events was
included in the benchmarked time.

A new benchmark instead compares the timestamps the Linux kernel
assigned to the events on the virtual input device and the events on the
virtual output device. This eliminates the overhead from the
benchmarking script, and it turns out that after eliminating that
overhead the latency turns out to be significantly lower than previously
thought.