Refactor Hook. Remove --hook withhold.
Split Hook into two parts: a Trigger that is responsible for keeping
track of when the Hook should trigger, and the rest of the Hook which is
responsible for executing the effects. The withhold flag does not
interact nicely with this setup and has therefore been removed in the
same commit. The withhold flag is planned to be replaced with a
--withhold argument in the future.