Shifted the responsibility of tracking activating_event to Hook.
I have mixed feelings about this refactor. On the one hand, it improves
the separtion of tasks: Triggers are also used by Withholds, and the
Withhold does not care about the activating_event, only the Hook does.
Therefore it makes sense that it is the Hook's duty to keep track of the
activating_event.

On the other hand, it is now possible for an invalid state to be
represented in the type system, i.e. to have a Hook where the Trigger is
active but the activating_event is None.

I believe that the clearer code from seperation of tasks makes this a
net win, but not a wide one.