Hide the --control-fifo argument behind a feature flag.
Rationale: the current implementation is half-baked and may need to be
changed in backwards incompatible ways. However, as it is high time to
release version 1.4, I will start to put features that have been worked
on but are not quite ready for release yet behind flags so they stop
causing even more delays for the next version release.