Lower the version number of the cache file format to 0.
This represents that the file format for chache is still in development.
To be clear, evsieve maintains backwards compatibility with released
versions, but makes no backwards compatibility guarantees about previous
commits on the main branch. Until a stable version of evsieve is
released, the file format for the cache is subject to be changed without
notice.