Add a rudimentary framework to replace for unittest/unittest.py in Rust.
The advantage of this new framework is that it does not create any actual I/O devices, does not require root to run, and is a lot faster.

The disadvantage is that due to not actually creating the I/O devices, it is not a holistic test. However, most tests in unittest/unittest.py merely serve to test how arguments get handled, and would therefore be more fit for this new framework.