Simplify the udev instructions.
The recommendation to give the user's account permission to read/write input devices has been removed due to security concerns regarding doing so, as it would allow any program to keylog and inject events for every other user. Although the user may decide that that is not a risk in their threat model, it is not appropriate to just recommend that.

The recommendation to use DynamicUser has been removed to keep things simple. Although it would be a sensible action and a good practice if somebody intends to, say, ship preconfigured evsieve configurations as part of a specialized Linux distro, it feels too complicated to be recommended to the average user. Adding a new system account and messing with the permissions on /dev/uinput feels like something that should only be done by somebody who actually understands what they are doing, and are not the kind of instructions that should be followed blindly from a guide.